+++ cpu/cpu_darwin_nocgo.go
@@ -1,0 +1,14 @@
+// +build darwin
+// +build !cgo
+
+package cpu
+
+import "github.com/shirou/gopsutil/common"
+
+func perCPUTimes() ([]CPUTimesStat, error) {
+	return []CPUTimesStat{}, common.NotImplementedError
+}
+
+func allCPUTimes() ([]CPUTimesStat, error) {
+	return []CPUTimesStat{}, common.NotImplementedError
+}
